+++ src/main/java/com/danyatheworst/location/dto/DeleteLocationRequestDto.java
@@ -5,13 +5,12 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
 public class DeleteLocationRequestDto {
-    private BigDecimal lat;
-    private BigDecimal lon;
+    private double lat;
+    private double lon;
 }
